pve-rs: fabrics: helper for mapping interface names
Add a helper method for proxmox-network-interface-pinning to rename
all interfaces of a node in the fabric config according to the mapping
generated by the pinning tool. This method renames all occurences of
an interface name to the pinned version. The logic has been taken and
reimplemented from the Perl helper itself.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250717152841.397830-3-s.hanreich@proxmox.com